Major tui refactor, featuring componentization and a more OO approach.
